+++ pages/list.jsx
@@ -1,23 +1,22 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import Head from 'next/head';
 
 import appContainer from '@containers/appContainer';
 
 import useSWR from 'swr';
 
-import TriviaListContainer from '../containers/TriviaListContainer';
 import TriviaCard from '../components/commons/TriviaCard';
 import TriviaModal from '../components/commons/TriviaModal';
 
 
 function ListPage() {
   const { apiGet } = appContainer.useContainer();
-  const { data, error } = useSWR('/trivias', () => apiGet('/trivias/user'));
+  const { data, error } = useSWR('/trivias/list', () => apiGet('/trivias/list', { page: 1 }));
 
   if (error) return <div>failed to load</div>;
   if (!data) return <div>loading...</div>;
 
-  return <div>hello!</div>;
+  const triviasList = data?.docs;
 
   return (
     <>
